+++ src/lib/summary.ts
@@ -1,12 +1,9 @@
-import fs from "fs";
 import { Matches } from "./comparison";
 import { HTMLSummaryFormatter } from "./htmlSummaryFormatter";
 import { JSONFormatter } from "./jsonFormatter";
 import { Range } from "./range";
-export type RangesTuple = [Range, Range];
+import { Clustered, Match, RangesTuple, Utils } from "./utils";
 
-export type Clustered<T> = T[][];
-export type Match = [string, string, RangesTuple[]];
 /**
  * @param minimumLinesInLargestFragment The minimum amount of lines required by the largest code fragment. The default
  * is 1.
@@ -21,83 +18,12 @@ export interface FilterOptions {
 }
 
 export class Summary {
-  public static optionsToString(optionsArray: Array<[string, string | number]>): string {
-    return (
-      optionsArray
-        .map(([flag, optionValue]) => {
-          if (typeof optionValue === "string" && optionValue.includes(" ")) {
-            return [flag, `'${optionValue}'`];
-          }
-          return [flag, optionValue];
-        })
-        .map(([flag, optionValue]) => `${flag} ${optionValue}`)
-        .join(" ") + "\n"
-    );
-  }
-
-  /**
-   * Counts the total amount of lines that correspond with the first and second file.
-   * @param rangesTupleArray The rangesTupleArray you want the lines count of.
-   * @returns A tuple that contais the line for the first and second file respectively.
-   */
-  public static countLinesInRanges(rangesTupleArray: RangesTuple[]): [number, number] {
-    return rangesTupleArray
-      .map(([range1, range2]) => [range1.getLineCount(), range2.getLineCount()] as [number, number])
-      .reduce(([acc1, acc2], [next1, next2]) => [acc1 + next1, acc2 + next2]);
-  }
-
-  /**
-   * Colours your text with the given colour. Only works for terminal output.
-   * @param colour The colour you want your text to be.
-   * @param text The text you want to colour.
-   * @param consoleColours If you want to colour the text or not.
-   */
-  public static colour(colour: string, text: string): string {
-    if (!Summary.colours.has(colour)) {
-      return text;
-    }
-    return `${Summary.colours.get(colour)}${text}${Summary.colours.get("Reset")}`;
-  }
-
-  public static getScoreForFiles(
-    matches: RangesTuple[],
-    matchedFile: string,
-    matchingFile: string,
-  ): [number, number] {
-    const [matchedLinesInMatchedFile, matchedLinesInMatchingFile]: [
-      number,
-      number,
-    ] = Summary.countLinesInRanges(matches);
-
-    const linesInMatchedFile: number = this.countLinesInFile(matchedFile);
-    const linesInMatchingFile: number = this.countLinesInFile(matchingFile);
-    const scoreMatchedFile: number = Math.round(
-      (matchedLinesInMatchedFile / linesInMatchedFile) * 100,
-    );
-    const scoreMatchingFile: number = Math.round(
-      (matchedLinesInMatchingFile / linesInMatchingFile) * 100,
-    );
-
-    return [scoreMatchedFile, scoreMatchingFile];
-  }
   private static readonly defaultFilterOptions: FilterOptions = Object.freeze({
     fragmentOutputLimit: undefined,
     minimumFragmentLength: 0,
   });
-  private static readonly colours: Map<string, string> = new Map([
-    ["FgRed", "\x1b[31m"],
-    ["FgGreen", "\x1b[32m"],
-    ["Reset", "\x1b[0m"],
-  ]);
-  // Maps the file to the amount of lines in that file.
-  private static readonly linesInFileMap: Map<string, number> = new Map();
 
-  private static countLinesInFile(fileName: string): number {
-    if (!this.linesInFileMap.has(fileName)) {
-      this.linesInFileMap.set(fileName, fs.readFileSync(fileName, "utf8").split("\n").length);
-    }
-    return this.linesInFileMap.get(fileName) as number;
-  }
+  private static readonly htmlFormatter: HTMLSummaryFormatter = new HTMLSummaryFormatter();
 
   public readonly gapSize: number;
   public readonly clusteredResults: Clustered<Match>;
@@ -209,14 +135,14 @@ export class Summary {
    * @param options The options that were used to generate this report.
    */
   public toHTML(comment?: string, options?: Array<[string, string | number]>): string {
-    return HTMLSummaryFormatter.format(this.toJSON(comment, options));
+    return Summary.htmlFormatter.format(this.toJSON(comment, options));
   }
   /**
    * Returns the maximum from the tuples returned from [[this.countLinesInRanges]].
    * @param rangesTupleArray The rangesTupleArray you want the maximum line count of.
    */
   public getMaxMatchingLineCount(rangesTupleArray: RangesTuple[]): number {
-    return Math.max(...Summary.countLinesInRanges(rangesTupleArray));
+    return Math.max(...Utils.countLinesInRanges(rangesTupleArray));
   }
 
   public toString(
@@ -233,13 +159,13 @@ export class Summary {
       output += comment + "\n";
     }
     if (optionsArray && optionsArray.length > 0) {
-      output += `Options: ${Summary.optionsToString(optionsArray)}\n`;
+      output += `Options: ${Utils.optionsToString(optionsArray)}\n`;
     }
 
     for (let index = 0; index < this.clusteredResults.length; index += 1) {
       let clusterNameString = `Cluster ${index + 1}\n`;
       if (consoleColours) {
-        clusterNameString = Summary.colour("FgRed", clusterNameString);
+        clusterNameString = Utils.colour("FgRed", clusterNameString);
       }
       output += clusterNameString;
       output += this.groupToString(this.clusteredResults[index], consoleColours).replace(
@@ -267,7 +193,7 @@ export class Summary {
       .map(match => JSON.stringify(match, JSONFormatter.JSONReplacerFunction))
       .join("\n\t\t");
 
-    const [scoreMatchedFile, scoreMatchingFile] = Summary.getScoreForFiles(
+    const [scoreMatchedFile, scoreMatchingFile] = Utils.getScoreForFiles(
       matches,
       matchedFile,
       matchingFile,
@@ -275,7 +201,7 @@ export class Summary {
 
     let returnString: string = `\t${matchedFile}(${scoreMatchedFile}%) + ${matchingFile}(${scoreMatchingFile}%) => \n`;
     if (consoleColours) {
-      returnString = Summary.colour("FgGreen", returnString);
+      returnString = Utils.colour("FgGreen", returnString);
     }
 
     return `${returnString}\t\t${matchesString}`;
